0030537: Visualization - wrapping text in font text formatter
Font_TextFormatter inherits Standard_Transient, now it is given as a handle in functions.
Graphic3d_Text - extended with Font_TextFormatter to be able to have it filled out of text render. If it is not defined here, the default text formatter of context is used.
OpenGl_Context - has default Font_TextFormatter for rendering OpenGl_Text.

# Conflicts:
#	src/Font/Font_TextFormatter.cxx
#	src/Font/Font_TextFormatter.hxx
#	src/Graphic3d/Graphic3d_Text.hxx
#	src/OpenGl/OpenGl_Context.cxx
#	src/OpenGl/OpenGl_Text.cxx
#	src/OpenGl/OpenGl_TextBuilder.cxx
#	src/StdPrs/StdPrs_BRepTextBuilder.cxx