Fix naming conflict between X11 headers and VTK 9.2.2+
Currently, the GLX headers include the X11 headers which use a
preprocessor define to make Status an int. However, VTK has a class
called Status, and this define replaces this class name with int which
results in compilation errors. This patch undefs Status and Success,
which are both defined in the X11 headers if they exist so that there
are no conflicts within the VTK headers for newer versions.